requirements.txt
openai==0.27.0
python-dotenv==1.0.0
django
Jinja2==3.1.2
pyyaml==6.0
psycopg2==2.9.6
pdfkit==1.0.0
python-decouple